+++ packages/docs/angular-docs/src/components/layout/app-docs-layout.tsx
@@ -24,12 +24,12 @@ import {
   DocsLayout,
   type DocsLayoutSettings,
 } from "@qualcomm-ui/react-mdx/docs-layout"
-import {SiteSearch} from "@qualcomm-ui/react-mdx/site-search"
 import {HeaderBar} from "@qualcomm-ui/react/header-bar"
 
 import angularPkg from "../../../../../frameworks/angular/package.json"
 
 import {DemoUpdateContextProvider} from "./demo-update-context"
+import {GlobalSearch} from "./global-search"
 import {ResetDemoDimensionsButton} from "./reset-demo-dimensions-button"
 import {useAngularDemoReloader} from "./use-angular-demo-reloader"
 import {useRouterLinkInterceptor} from "./use-router-link-interceptor"
@@ -94,7 +94,7 @@ export function AppDocsLayout({children, ...props}: Props): ReactNode {
             </HeaderBar.Nav>
 
             <HeaderBar.ActionBar>
-              <SiteSearch />
+              <GlobalSearch />
               <ResetDemoDimensionsButton />
               <HeaderRepositoryLink />
               <GlobalConfig />