+++ packages/frameworks/react-core/vitest.config.ts
@@ -1,19 +1,12 @@
-import babel from "@rolldown/plugin-babel"
 import tailwindcss from "@tailwindcss/vite"
-import react, {reactCompilerPreset} from "@vitejs/plugin-react"
+import react from "@vitejs/plugin-react"
 import {defineConfig} from "vitest/config"
 
 import {getReactTestConfig} from "@qualcomm-ui/react-test-utils"
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [
-    tailwindcss(),
-    react(),
-    babel({
-      presets: [reactCompilerPreset()],
-    }),
-  ],
+  plugins: [tailwindcss(), react({compiler: true})],
   resolve: {
     tsconfigPaths: true,
   },